Vostok/requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-materializecss-form==1.1.14
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.3
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
django-allauth
sqlparse==0.3.0
urllib3==1.25.6
django-filter==2.1.0
djangorestframework==3.9.3
Markdown==3.1